extract the db reset grpc call to another file
